Oh yea, it's the Sheep Shear Club: thanks to part 1, it's already possible to buy individual products. Now, click "Pricing". The real point of the site is to get a subscription that'll send your awesome sheep awesome shearing accessories monthly. After vigorous meetings with our investors, we've decided to offer 2 plans: the Farmer Brent at $99/month and the New Zealander at $199/month. But other than this fancy looking page, we haven't coded anything for this yet.
Tell Stripe about the Plans

Our first step is to tell Stripe about these two plans. To do that, open your trusty Stripe dashboard, make sure you're in the test environment, and find "Plans" on the left. We're going to create the two plans in Stripe by hand. The ID can be any unique string and we'll use this forever after to refer to the plan. Use, farmer_brent_monthly. Then, fill in a name - that's less important and set the price as $99 per month. Create that plan!

And yes, later, we'll add yearly versions of each plan... but one step at a time! Repeat that whole process for the New Zealander: set its ID to new_zealander_monthly, give it a name and set its price and interval. Perfect!

Manage the Plans in Our Code

Now, head back to our code and open the src/Subscription directory. I added two classes to help us stay organized. The first is called SubscriptionHelper:

 33 lines  src/Subscription/SubscriptionHelper.php
namespace App\Subscription;
class SubscriptionHelper
{
    /** @var SubscriptionPlan[] */
    private $plans = [];
    public function __construct()
    {
        // todo - add the plans
//        $this->plans[] = new SubscriptionPlan(
//            'STRIPE_PLAN_KEY',
//            'OUR PLAN NAME',
//            'PRICE'
//        );
    }
    /**
     * @param $planId
     * @return SubscriptionPlan|null
     */
    public function findPlan($planId)
    {
        foreach ($this->plans as $plan) {
            if ($plan->getPlanId() == $planId) {
                return $plan;
            }
        }
    }
}
And as you can see... it's not very helpful... yet. But pretty soon, it will keep track of the two plans we just added. We'll use the second class - SubscriptionPlan:

 34 lines  src/Subscription/SubscriptionPlan.php
namespace App\Subscription;
class SubscriptionPlan
{
    private $planId;
    private $name;
    private $price;
    public function __construct($planId, $name, $price)
    {
        $this->planId = $planId;
        $this->name = $name;
        $this->price = $price;
    }
    public function getPlanId()
    {
        return $this->planId;
    }
    public function getName()
    {
        return $this->name;
    }
    public function getPrice()
    {
        return $this->price;
    }
}
To hold the data for each plan: the plan ID, name and price. But we won't save these to the database.