Updating the Database
Since our goal is to update the database during checkout, go back to OrderController and find the chargeCustomer() method that holds all the magic.

But instead of putting the code to update the database right here, let's add it to SubscriptionHelper: this class will do all the work related to subscriptions. Add a new method at the bottom called public function addSubscriptionToUser() with two arguments: the \Stripe\Subscription object that was just created and the User that the Subscription should belong to

Inside, start with $subscription = $user->getSubscription(). So, the user may already have a row in the subscription table from a previous, expired subscription. If they do, we'll just update that row instead of creating a second row. Every User will have a maximum of one related row in the subscription table. It keeps things simple.

But if they don't have a previous subscription, let's create one: $subscription = new Subscription(). Then, $subscription->setUser($user)

Our other todo is to update the fields on the Subscription object: $stripeSubscriptionId and $stripePlanId. To keep things clean, open Subscription and add a new method at the bottom: public function activateSubscription() with two arguments: the $stripePlanId and $stripeSubscriptionId. Set each of these onto the corresponding properties. Also add $this->endsAt = null. We'll talk more about that later, but this field will help us know whether or not a subscription has been cancelled.