Finally, call that method! $this->subscriptionHelper->updateCardDetails() and pass it $user and $stripeCustomer
No matter how you checkout, we're going to make sure your card details are updated in our database!

Before we try it out and prove how awesome we are, I want to add one more thing: I want to be able to tell the user when they will be billed next.