When the user cancels, we need to somehow update the user's row in the subscription table so that we know this happened! And actually, it's kind of complicated: the user canceled, but the subscription should still be active until the end of the month. Then it'll really be canceled. So, how the heck can we manage this?
Here's the plan: we are not going to delete the subscription from the subscription table, because it's still active until the period end. Instead, we'll set the endsAt date field to when the subscription will expire

That way, we'll know if the subscription is still active, meaning it's before the endsAt date, or it's fully canceled, because it's after the endsAt date.

At the bottom of Subscription, add a helper function to do this: public function deactivateSubscription()

Since we know the user has paid through the end of the period, we can use that: $this->endsAt = $this->billingPeriodEndsAt. Also set $this->billingPeriodsEndsAt = null - just so we know that there won't be another bill at the end of this month

Cool! To deactivate the subscription in the controller, it's as easy as saying $subscription->deactivateSubscription() and then saving it to the database with the standard persist() and flush() Doctrine code

And that should do it! Let's give this guy a try. Go to the account page, then press the new "Cancel Subscription" button. Ok, looks good! Check the customer page in the Stripe dashboard. Yes! The most recent subscription - the one we're dealing with in our code - is active but will cancel at the end of the month.

But if you look at the Account page, everything here still looks "Active". We updated the endsAt field on the subscription, but our code in this template isn't smart enough... yet.