Open account.html.twig. Hmm, I need an easy way to know whether or not a subscription is active, and if it is active, whether or not it's in this canceled state.
To help with this, let's create two methods inside the Subscription class. First, public function isActive()
Meaning: does the user still have an active subscription, even if it will cancel at the month's end? So, if $this->endsAt === null, then the subscription is definitely active. OR, $this->endsAt is greater than right now, new \DateTime()
Meaning the subscription is canceled, but is ending in the future.
The second method we need is public function isCanceled()
Meaning: if the subscription is active, has the user actually canceled it or not? This will simply be, return $this->endsAt !== null

Oh man, our setup is getting fancy! Let's get even fancier with one more helper method, this time in User. Add a new public function hasActiveSubscription()

A User has an active subscription if they have a subscription object related to them and that subscription object isActive(). That'll save us some typing whenever we need to check whether or not a user has an active subscription.

Making the Account Template Awesome

Ok, back to the account template! This time, to be heros!

First, that "Cancel Subscription" button should only be there if the user has an active subscription. No problem! Add if app.user.hasActiveSubscription()
But even here, if the user has already canceled their subscription, we don't want to keep showing them this button. Add another if: if app.user.subscription.isCancelled()
Then add a little "TODO" to add a re-activate button. If they've cancelled, they might remember how cool your service is and want to come back LATER and reactivate! In the else, show them the Cancel button.

Finish up the endif and the other endif. And actually, copy these first two lines: we need to re-use them further below. In the section that tells us whether or not we have an active subscription, we now have three states: "active", "active but canceled," and "none." Replace the old if statement with the two that you just copied. If the subscription is canceled, add label-warning and say "Canceled". Else, we know it's active
If a user doesn't have any type of active subscription, keep the "none" from before.

Finally, copy just the first if statement and scroll down to "Next Billing at". We should only show the next billing period if the user has an active subscription, not just if they have a related subscription object, because it could be canceled. Paste the if statement over this one
Finally, do the same thing down below for the credit card: I don't want to confuse someone by showing them credit card information when they don't have a subscription

Phew! Ok, refresh! It's beautiful! There's our todo for the reactivate button and the subscription is canceled. But wait! We don't want the "Next Billing at" and credit card information to show up.