Reactivate the subscription
In the Stripe API docs, under the "canceling" section, there's actually a spot about reactivating canceled subscriptions, and it's really interesting! It says that if you use the at_period_end method of canceling, and the subscription has not yet reached the period end, then reactivating is easy: just set the subscription's plan to the same plan ID that it had originally. Internally, Stripe knows that means I want to not cancel the subscription anymore.

Route and Controller Setup

Let's hook it up! We're going to need a new endpoint that reactivates a subscription. In ProfileController, add a new public function reactivateSubscriptionAction(). Give it a route set to /profile/subscription/reactivate and a name: account_subscription_reactivate

Good start! With this in place, copy the route name, open account.html.twig and go up to the "TODO" we added a few minutes ago. Paste the route, just to stash it somewhere, then copy the entire cancel form and put it here. Update the form action with the new route name, change the text, and use btn-success to make this look like a really happy thing

Refresh and enjoy the nice, new Reactivate Subscription button. Beautiful!