The update card form (Re-using the Card Form)
Eventually, our customers will need to update the credit card that we have stored. And on surface, this is pretty easy: the card is just a property on the Stripe customer called source. And basically, we need to nearly duplicate the checkout process: show the user a card form, exchange their card info for a Stripe token, submit that token, and then save it on the user as their new card.

Ok, let's rock! Step 1: add an "Update Card" button to the account page and make it re-use the same card form we already built for checkout. Because, ya know, reusing code is awesome!

Re-using the Card Form

In account.html.twig, find the credit card section. Add a button, give it some classes for styling and a js-open-credit-card-form class.
Next, find the col-xs-6. This is the right side of the page, and I've kept it empty until now just so we can show the form here. Add a div with a JS class so we can hide/show this element: js-update-card-wrapper. Make it display: none; by default. Ok, I want to re-use the entire checkout form right here. Fortunately, we already isolated that into its own template: _cardForm.html.twig. Yay!

In account.html.twig, use the Twig include() function to bring that in. Ok! Let's hide/show this form whenever the user clicks the "Update Card" button. At the top of the file, override the block called javascripts, and call endblock. Inside, call the parent() function. In this project, any JS we put here will be included on the page.

Add a script tag and a very simple document.ready() block. Inside of that, find the .js-open-credit-card-form element and on click, create a callback function. Start with the normal e.preventDefault(). Now, find the other wrapper element, which is js-update-card-wrapper. Call slideToggle() on that to show/hide it

So, fairly easy stuff.

Well, maybe we should see if it works first. Refresh! Ah, it doesn't! Huge error:

  Variable "error" does not exist in _cardForm.html.twig at line 56
Hmm, checkout that template:

 68 lines  templates/views/order/_cardForm.html.twig
<form action="" method="POST" class="js-checkout-form checkout-form">
    <div class="row">
        <div class="col-xs-8 col-sm-6 col-sm-offset-2 text-center">
            <div class="alert alert-danger js-checkout-error {{ error ? '' : 'hidden' }}">{{ error }}</div>
        </div>
    </div>
</form>
Ah yes, on the checkout page, after we submit, if there was an error, we set this variable and render it here. For now, we don't have any errors.