Update credit card information endpoint.
Open ProfileController and add a new endpoint: public function updateCreditCardAction(). Give it the URL /profile/card/update and a fancy name: account_update_credit_card. Add methods={"POST"} to be extra cool

With this in place, we need to update the form to submit here. Check out _cardForm.html.twig. Hmm, the action attribute is empty:

 68 lines  templates/views/order/_cardForm.html.twig
<form action="" method="POST" class="js-checkout-form checkout-form">
</form>
That's because we want the checkout form to submit right back to /checkout. But this won't work for the update card form: it should submit to a different URL.
Now we can override this! In account.html.twig, add another variable to the include: formAction set to path() and the new route name.

Refresh and check out the source. Ok, the form action is ready!

Instead, render a new variable called formAction and pipe that to the default filter with empty quotes