Saving credit card information and updating cardLast4 and cardBrand
Saving credit card information

Let's get to work in ProfileController! But actually... this will be very similar to our checkout logic, so let's go steal code! Copy the line that fetches the stripeToken POST parameter and then head back to ProfileController. Make sure you have a Request argument and the Request use statement

Then, first, paste that line to fetch the token. Second, fetch the current user object with $user = $this->getUser(). And third, we need to make an API request to stripe that updates the Customer and attaches the token as their new card. That means we'll be using the StripeClient. Fetch it first with $stripeClient = $this->stripeClient. Here's the awesome part: open StripeClient. We already have a method called updateCustomerCard()

We pass the User object and the submitted payment token and it sets it on the Customer and saves.

Victory for code organization! In the controller, just say $stripeClient->updateCustomerCard() and pass it $user and $token

That takes care of things inside Stripe.

Updating cardLast4 and cardBrand

Now, what about our database? Do we store any information about the credit card? Actually, we do! In the User class, we store cardLast4 and cardBrand. With the new card, this stuff probably changed!

But wait, we've got this handled too guys! Open SubscriptionHelper and check out the handy updateCardDetails() method Just pass it the User and \Stripe\Customer and it'll make sure those fields are set.

In ProfileController, call this: ->subscriptionHelper->updateCardDetails() passing  and ... which doesn't exist yet. Fortunately, updateCustomerCard() returns that, so create that variable on that line. That's it! Add a success message so that everyone feels happy and joyful, and redirect back to the profile page.

Time to try it! Refresh and put in the fake credit card info. But use a different expiration: 10/25. Hit "Update Card". Ok, it looks like it worked. Refresh the Customer page in Stripe. The expiration was 10/20 and now it's 10/25. Card update successful!

But, we still need to handle one more case: when the card update fails.