There are actually two ways for the credit card update to fail. Most failures happen immediately, and are handle via JavaScript. But a few don't happen until we try to attach the card to the customer.
Let's see an example. In Stripe's documentation, find the "Testing" section - it's under the Payments header in the new design. Down the page a bit, you'll find a table full of cards that will work or fail for different reasons. Find the one that ends in 0002 and copy it. Fill the form out using this and update the card.

Ah, 500 error!

Your card was declined

Ok, a \Stripe\Error\Card exception was thrown the moment that we tried to save the new customer card back to Stripe. We did handle this situation on our checkout page, so we just need to also handle it here.