Let's get right to work on our webhook endpoint.
In the src/Controller directory, create a new WebhookController class. Make it extend a AbstractController class I created - that just has a few small shortcuts

Now, create the endpoint with public function stripeWebhookAction(). Give an @route annotation set to /webhooks/stripe and a similar name. Make sure you have the Route use statement

Start simple: return a new Response() from the HttpFoundation component
That's just enough to try it out: find your browser and go to /webhooks/stripe. It's alive!

Decoding the Event

Thanks to RequestBin, we know more or less what the JSON body will look like. The most important thing is this event id. Let's decode the JSON and grab this.

To do that, add $data = json_decode(), but pause there. We need to pass this the body of the Request. In Symfony, we get this by adding a Request argument - don't forget the use statement! Then, use $request->getContent(). Also, pass true as the second argument so that json_decode returns an associative array

Next, it shouldn't happen, but just in case, if  is null, that means Stripe sent us invalid JSON. Shame on you Stripe! Throw an exception in this case... and make sure you spell Exception correctly! Finally, get the  from $data['id']

Ok, let's refocus on the next steps. Ultimately, I want to read these fields in the event, find the Subscription in the database, and cancel it. But instead of reading the JSON body directly, we're going to use Stripe's API to fetch the Event object by using this .

Wait, but won't that just return the exact same data we already have? Yes! We do this not because we need to, but for security. If we read the request JSON directly, it's possible that the request is coming from some external, mean-spirited person instead of from Stripe. By fetching a fresh event from Stripe, it guarantees the event data is legitimate.