Cancel the subscription
First, we need to find the Subscription in our database. And check this out: the subscription id lives at data.object.id. That's because this type of event embeds the subscription in question. Other event types will embed different data.

Add $stripeSubscriptionId = $stripeEvent->data->object->id

Next, the subscription table has a stripeSubscriptionId field on it. Let's query on this! Because I already know I'll want to re-use this next code, I'll put the logic into a private function. On this line, call that future function with $subscription = $this->findSubscription() and pass it $stripeSubscriptionId

Scroll down and create this: private function findSubscription() with its $stripeSubscriptionId argument

Query by adding $subscription = $this->getDoctrine()->getRepository('App:Subscription') and then findOneBy() passing this an array with one item: stripeSubscriptionId - the field name to query on - set to $stripeSubscriptionId

If there is no matching Subscription... well, that shouldn't happen! But just in case, throw a new Exception with a really confused message. Something is not right.

Finally, return the $subscription on the bottom

Ok, head back up to the action method. Hmm, so all we really need to do now is call the cancel() method on ! But let's get a little bit more organized. Open SubscriptionHelper and add a new method there: public function fullyCancelSubscription() with the Subscription object that should be canceled. Below, really simple, say $subscription->cancel(). Then, use the Doctrine entity manager to save this to the database

Mind blown!

Back in the controller, call this! $this->subscriptionHelper->fullyCancelSubscription($subscription)

And that is it!

Yep, there's some setup to get the webhook controller started, but now we're in really good shape.

Of course... we have no way to test this... So, ya know, just make sure you do a really good job of coding and hope for the best! No, that's crazy!, I'll show you a few ways to test this next. But also, don't forget to configure your webhook URL in Stripe once you finally deploy this to beta and production. I have a webhook setup for each instance on KnpU.