Using your Local Machine with Ngrok
The second option is to point a webhook at your local development machine. But wait! That's not possible: our local machine is not accessible by the internet.

Well... that doesn't have to be true. By using a cool utility called Ngrok, you can temporarily tunnel a public URL to your computer.

Let's try it! Since I already have ngrok installed,
To use https I must run first ngrok authtoken xxxxxxxxxxxxxxxxxxxxxxxX_7xxxxxxxxxxxxxxxxxx
where authtoken can be obtain from ngrok site
I can use it from any directory on my system:
 ngrok http https://localhost:8000

That will expose port 8000 - the one we're serving our site on - to the web via this cool public URL.

Using the Ngrok URL as a Webhook Endpoint

Now we're super dangerous! In your Stripe Webhook configuration, add an endpoint. Paste the URL and put this in the "Test" environment. For now, just receive the customer.subscription.deleted event. Create the endpoint!

Oh, wait, make sure the endpoint URL ends in /webhooks/stripe. That's better!

In our app, we already have an active subscription. So, in Stripe, click "Customers" and open our one Customer. Find the top subscription and... cancel it! Immediately!

That should cause a webhook to be sent to our local machine. So, moment of truth: refresh the account page! Oh no! The subscription doesn't look canceled!

Hmm, go back to Stripe. At the bottom of the Customer page, you can see all the events for this Customer. Or, another way to look at this is by clicking "Events & webhooks" on the left. Ah! And we can see the customer.subscription.deleted event! And at the bottom, it shows that the webhook to our ngrok URL was successful.

So, refresh the account page again. Ah, now the subscription is canceled. We were just too fast the first time.

So, choose your favorite method of testing these crazy webhook things, and make sure they are bug-free.